Fix broken conversion of 16 bit interleaved stereo to internal format.
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@12441 a1c6a512-1295-4272-9138-f99709370657

+++ apps/dsp.c
@@ -191,7 +191,7 @@ static int convert_lte_native_interleaved_stereo(
 
     do
     {
-        short slr = *s++;
+        int32_t slr = *s++;
 #ifdef ROCKBOX_LITTLE_ENDIAN
         *dl++ = (slr >> 16) << scale;
         *dr++ = (int32_t)(short)slr << scale;
